Optimization on editor display and logic

+++ Editor/Main/MagicLinksScriptsGenerator.cs
@@ -1,7 +1,6 @@
 #if UNITY_EDITOR
 using System.Collections.Generic;
 using System.IO;
-using System.Text.RegularExpressions;
 using UnityEngine;
 using UnityEditor;
 
@@ -9,10 +8,18 @@ namespace MagicLinks
 {
     public static class MagicLinksScriptsGenerator
     {
+        private static string _cachedTemplate;
+
+        private static string GetTemplate()
+        {
+            if (_cachedTemplate != null) return _cachedTemplate;
+            _cachedTemplate = File.ReadAllText(
+                MagicLinksUtilities.GetPackageRelativePath(MagicLinksConst.MagicVariablesTemplate));
+            return _cachedTemplate;
+        }
+
         public static void GenerateMagicVariablesScript(bool ifMissing)
         {
-            string classContent = GetMagicVariablesScriptContent();
-
             MagicLinksUtilities.CreateVariablesFolder();
 
             string newClassPath = Path.Combine(MagicLinksConst.ConfigurationPath,
@@ -20,83 +27,93 @@ namespace MagicLinks
 
             if (ifMissing && File.Exists(newClassPath)) return;
 
+            string classContent = GetMagicVariablesScriptContent();
+
+            if (File.Exists(newClassPath) && File.ReadAllText(newClassPath) == classContent) return;
+
             File.WriteAllText(newClassPath, classContent);
 
-            AssetDatabase.Refresh();
+            AssetDatabase.ImportAsset(newClassPath);
         }
 
         public static string GetMagicVariablesScriptContent()
         {
-            string classContent =
-                File.ReadAllText(MagicLinksUtilities.GetPackageRelativePath(MagicLinksConst.MagicVariablesTemplate));
+            var allTypes = MagicLinksUtilities.GetAllTypes();
 
-            classContent = classContent.Replace("MagicVariablesTemplate", MagicLinksConst.MagicVariableClassName);
+            var sb = new System.Text.StringBuilder(GetTemplate());
 
-            string variables = string.Empty;
+            sb.Replace("MagicVariablesTemplate", MagicLinksConst.MagicVariableClassName);
 
-            foreach (string customType in MagicLinksUtilities.GetAllTypes())
+            var variables = new System.Text.StringBuilder();
+            foreach (string customType in allTypes)
             {
-                variables += GetDict(MagicLinksConst.VariableDictTemplate, customType, customType.ToUpper());
-                variables += GetDict(MagicLinksConst.ListVariableDictTemplate, customType, customType.ToUpper() + MagicLinksConst.ListDict);
+                AppendDict(variables, MagicLinksConst.VariableDictTemplate, customType, customType.ToUpper());
+                AppendDict(variables, MagicLinksConst.ListVariableDictTemplate, customType, customType.ToUpper() + MagicLinksConst.ListDict);
             }
 
-            variables += "\n \n";
+            variables.Append("\n \n");
 
-            foreach (string customType in MagicLinksUtilities.GetAllTypes())
+            foreach (string customType in allTypes)
             {
-                variables += GetDict(MagicLinksConst.EventDictTemplate, customType,
+                AppendDict(variables, MagicLinksConst.EventDictTemplate, customType,
                     customType.ToUpper() + MagicLinksConst.EventDict);
-                variables += GetDict(MagicLinksConst.ListEventDictTemplate, customType,
+                AppendDict(variables, MagicLinksConst.ListEventDictTemplate, customType,
                     customType.ToUpper() + MagicLinksConst.ListDict + MagicLinksConst.EventDict);
             }
 
-            variables += MagicLinksConst.EventVoidDictTemplate;
+            variables.Append(MagicLinksConst.EventVoidDictTemplate);
 
-            classContent = classContent.Replace("//VARIABLESLISTS", variables);
-            classContent = classContent.Replace("/*", string.Empty);
-            classContent = classContent.Replace("*/", string.Empty);
-            classContent = classContent.Replace("#if UNITY_EDITOR", string.Empty);
-            classContent = classContent.Replace("#endif", string.Empty);
+            sb.Replace("//VARIABLESLISTS", variables.ToString());
+            sb.Replace("/*", string.Empty);
+            sb.Replace("*/", string.Empty);
+            sb.Replace("#if UNITY_EDITOR", string.Empty);
+            sb.Replace("#endif", string.Empty);
 
-            classContent = classContent.Replace("//STARTUSINGEDITOR", "#if UNITY_EDITOR");
-            classContent = classContent.Replace("//ENDUSINGEDITOR", "#endif");
+            sb.Replace("//STARTUSINGEDITOR", "#if UNITY_EDITOR");
+            sb.Replace("//ENDUSINGEDITOR", "#endif");
 
             //Generate Getters
-            string variablesGetter = string.Empty;
-            foreach (string customType in MagicLinksUtilities.GetAllTypes())
+            var variablesGetter = new System.Text.StringBuilder();
+            foreach (string customType in allTypes)
             {
-                variablesGetter += GetDict(MagicLinksConst.VariableGetterTemplate, customType, customType.ToUpper());
-                variablesGetter += GetDict(MagicLinksConst.ListVariableGetterTemplate, customType, customType.ToUpper() + MagicLinksConst.ListDict);
+                AppendDict(variablesGetter, MagicLinksConst.VariableGetterTemplate, customType, customType.ToUpper());
+                AppendDict(variablesGetter, MagicLinksConst.ListVariableGetterTemplate, customType, customType.ToUpper() + MagicLinksConst.ListDict);
             }
 
-            classContent = classContent.Replace("//MAGICVARIABLESGETTER", variablesGetter);
+            sb.Replace("//MAGICVARIABLESGETTER", variablesGetter.ToString());
 
-            string eventsGetter = string.Empty;
-            foreach (string customType in MagicLinksUtilities.GetAllTypes())
+            var eventsGetter = new System.Text.StringBuilder();
+            foreach (string customType in allTypes)
             {
                 string eventGetter = GetDict(MagicLinksConst.EventGetterTemplate, customType,
                     customType.ToUpper() + MagicLinksConst.EventDict);
 
                 eventGetter = eventGetter.Replace("SHORT", customType.ToUpper());
 
-                var regex = new Regex(Regex.Escape(MagicLinksConst.EventDict));
-                eventsGetter = regex.Replace(eventsGetter, MagicLinksConst.EventDict, 1);
-
-                eventsGetter += eventGetter;
+                eventsGetter.Append(eventGetter);
 
                 string listEventGetter = GetDict(MagicLinksConst.ListEventGetterTemplate, customType,
                     customType.ToUpper() + MagicLinksConst.ListDict + MagicLinksConst.EventDict);
 
                 listEventGetter = listEventGetter.Replace("SHORT", customType.ToUpper() + MagicLinksConst.ListDict);
 
-                eventsGetter += listEventGetter;
+                eventsGetter.Append(listEventGetter);
             }
 
-            eventsGetter += MagicLinksConst.EventVoidGetterTemplate;
+            eventsGetter.Append(MagicLinksConst.EventVoidGetterTemplate);
 
-            classContent = classContent.Replace("//MAGICEVENTSGETTER", eventsGetter);
+            sb.Replace("//MAGICEVENTSGETTER", eventsGetter.ToString());
 
-            return classContent.Split(MagicLinksConst.TemplateListenerSeparation)[0];
+            return sb.ToString().Split(MagicLinksConst.TemplateListenerSeparation)[0];
+        }
+
+        private static void AppendDict(System.Text.StringBuilder target, string dict, string t, string n)
+        {
+            int start = target.Length;
+            target.Append(dict);
+            target.Replace("TYPE", t, start, target.Length - start);
+            target.Replace("NAME", n, start, target.Length - start);
+            target.Append('\n');
         }
 
         public static string GetDict(string dict, string t, string n)
@@ -114,85 +131,117 @@ namespace MagicLinks
         {
             MagicLinksUtilities.CreateListenersFolder();
 
-            List<string> eventsListenersScriptsName = new List<string>();
-            List<string> variablesListenersScriptsName = new List<string>();
+            var allTypes = MagicLinksUtilities.GetAllTypes();
 
-            foreach (string t in MagicLinksUtilities.GetAllTypes())
+            HashSet<string> eventsListenersScriptsName = new HashSet<string>();
+            HashSet<string> variablesListenersScriptsName = new HashSet<string>();
+
+            foreach (string t in allTypes)
             {
                 eventsListenersScriptsName.Add(MagicLinksUtilities.GetEventListenerName(t) + ".cs");
-            }
-
-            string[] eventsFiles = Directory.GetFiles(MagicLinksConst.EventsListenersPath);
-
-            foreach (string file in eventsFiles)
-            {
-                CheckDeleteFile(file, everything, eventsListenersScriptsName);
-            }
-            
-            //-------------------------------------
-            
-            foreach (string t in MagicLinksUtilities.GetAllTypes())
-            {
                 variablesListenersScriptsName.Add(MagicLinksUtilities.GetVariableListenerName(t) + ".cs");
             }
 
-            string[] variablesFiles = Directory.GetFiles(MagicLinksConst.VariablesListenersPath);
+            bool anyDeleted = false;
 
-            foreach (string file in variablesFiles)
+            AssetDatabase.StartAssetEditing();
+            try
             {
-                CheckDeleteFile(file, everything, variablesListenersScriptsName);
+                foreach (string file in Directory.GetFiles(MagicLinksConst.EventsListenersPath))
+                {
+                    if (CheckDeleteFile(file, everything, eventsListenersScriptsName)) anyDeleted = true;
+                }
+
+                foreach (string file in Directory.GetFiles(MagicLinksConst.VariablesListenersPath))
+                {
+                    if (CheckDeleteFile(file, everything, variablesListenersScriptsName)) anyDeleted = true;
+                }
+            }
+            finally
+            {
+                AssetDatabase.StopAssetEditing();
             }
 
-            AssetDatabase.Refresh();
+            if (anyDeleted) AssetDatabase.Refresh();
         }
 
-        public static void CheckDeleteFile(string file, bool everything, List<string> toCompare) 
+        public static bool CheckDeleteFile(string file, bool everything, HashSet<string> toCompare)
         {
-            if (Path.GetExtension(file) != ".cs") return;
+            if (Path.GetExtension(file) != ".cs") return false;
 
             string fileName = Path.GetFileName(file);
-                
-            if(everything == false && fileName == MagicLinksUtilities.GetEventListenerName("Void") + ".cs") return;
 
-            if (everything == false && toCompare.Contains(fileName)) return;
+            if (everything == false && fileName == MagicLinksUtilities.GetEventListenerName("Void") + ".cs") return false;
+
+            if (everything == false && toCompare.Contains(fileName)) return false;
 
             File.Delete(file);
+            string metaPath = file + ".meta";
+            if (File.Exists(metaPath)) File.Delete(metaPath);
+            return true;
         }
 
         public static void GenerateListenersScripts()
         {
+            GenerateListenersScripts(false);
+        }
+
+        public static void GenerateListenersScripts(bool ifMissing)
+        {
             MagicLinksUtilities.CreateListenersFolder();
 
-            string template = File.ReadAllText(MagicLinksUtilities.GetPackageRelativePath(MagicLinksConst.MagicVariablesTemplate));
+            string template = GetTemplate();
             template = template.Replace("/*", string.Empty);
             template = template.Replace("*/", string.Empty);
 
             string[] separated = template.Split(MagicLinksConst.TemplateListenerSeparation);
-            
-            foreach (string t in MagicLinksUtilities.GetAllTypes())
+
+            bool anyWritten = false;
+
+            AssetDatabase.StartAssetEditing();
+            try
             {
-                //Event Listeners
-                string eventClassName = MagicLinksUtilities.GetEventListenerName(t);
+                foreach (string t in MagicLinksUtilities.GetAllTypes())
+                {
+                    //Event Listeners
+                    string eventClassName = MagicLinksUtilities.GetEventListenerName(t);
+                    string eventListenerPath = Path.Combine(MagicLinksConst.EventsListenersPath, eventClassName + ".cs");
+                    if (WriteIfChanged(eventListenerPath,
+                            CreateListenerContent(separated[1], eventClassName, t, "MagicEvents", "OnEventRaised"),
+                            ifMissing))
+                        anyWritten = true;
 
-                string eventListenerPath = Path.Combine(MagicLinksConst.EventsListenersPath, eventClassName + ".cs");
-                string eventListenerContent = CreateListenerContent(separated[1], eventClassName, t, "MagicEvents", "OnEventRaised");
-                File.WriteAllText(eventListenerPath, eventListenerContent);
-                
-                //Variable Listeners
-                string variableClassName = MagicLinksUtilities.GetVariableListenerName(t);
+                    //Variable Listeners
+                    string variableClassName = MagicLinksUtilities.GetVariableListenerName(t);
+                    string variableListenerPath = Path.Combine(MagicLinksConst.VariablesListenersPath, variableClassName + ".cs");
+                    if (WriteIfChanged(variableListenerPath,
+                            CreateListenerContent(separated[1], variableClassName, t, "MagicVariables", "OnValueChanged"),
+                            ifMissing))
+                        anyWritten = true;
+                }
 
-                string variableListenerPath = Path.Combine(MagicLinksConst.VariablesListenersPath, variableClassName + ".cs");
-                string variableListenerContent = CreateListenerContent(separated[1], variableClassName, t, "MagicVariables", "OnValueChanged");
-                File.WriteAllText(variableListenerPath, variableListenerContent);
+                //Event Listener VOID
+                string voidClassName = MagicLinksUtilities.GetEventListenerName("Void");
+                string voidListenerPath = Path.Combine(MagicLinksConst.EventsListenersPath, voidClassName + ".cs");
+                if (WriteIfChanged(voidListenerPath, separated[2], ifMissing)) anyWritten = true;
             }
-            
-            //Event Listener VOID
-            string voidClassName = MagicLinksUtilities.GetEventListenerName("Void");
-            string voidListenerPath = Path.Combine(MagicLinksConst.EventsListenersPath, voidClassName + ".cs");
+            finally
+            {
+                AssetDatabase.StopAssetEditing();
+            }
 
-            File.WriteAllText(voidListenerPath, separated[2]);
+            if (anyWritten) AssetDatabase.Refresh();
+        }
 
-            AssetDatabase.Refresh();
+        private static bool WriteIfChanged(string path, string content, bool ifMissing)
+        {
+            if (File.Exists(path))
+            {
+                if (ifMissing) return false;
+                if (File.ReadAllText(path) == content) return false;
+            }
+            File.WriteAllText(path, content);
+            return true;
         }
 
         public static string CreateListenerContent(string template, string n, string t, string link, string kind)
